+++ src/type-system.ts
@@ -30,6 +30,7 @@ import { fullFormats } from './formats'
 import type { CookieOptions } from './cookies'
 import { ValidationError } from './error'
 import type { MaybeArray } from './types'
+import { hasTransform } from './compose'
 
 const isISO8601 =
 	/(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d\.\d+([+-][0-2]\d:[0-5]\d|Z))|(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z))|(\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d([+-][0-2]\d:[0-5]\d|Z))/
@@ -619,7 +620,8 @@ export const ElysiaType = {
 
 				if (typeof value === 'string') return decode(value)
 
-				throw new ValidationError('property', schema, value)
+				// Is probably transformed, unable to check schema
+				return value
 			})
 			.Encode((value) => {
 				if (typeof value === 'string')